general tidying up, using uint_fast*_t datatypes, etc.

+++ src/shmem/include/ptl_internal_nit.h
@@ -22,6 +22,6 @@ extern ptl_ni_limits_t    nit_limits[4];
 int  PtlInternalNIValidator(const ptl_internal_handle_converter_t ni);
 void PtlInternalDeallocUnexpectedHeader(ptl_internal_buffered_header_t *const hdr);
 
-ptl_internal_buffered_header_t *PtlInternalAllocUnexpectedHeader(const unsigned int ni);
+ptl_internal_buffered_header_t *PtlInternalAllocUnexpectedHeader(const uint_fast8_t ni);
 #endif /* ifndef PTL_INTERNAL_NIT_H */
 /* vim:set expandtab: */
